+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Use latest csv-detective version [#77](https://github.com/etalab/udata-hydra/pull/77)
 - Compare content type / length to check if changed [#78](https://github.com/etalab/udata-hydra/pull/78) [#79](https://github.com/etalab/udata-hydra/pull/79)
 - Create a list of exceptions to analyse despite larger size [#85](https://github.com/etalab/udata-hydra/pull/85)
+- Enable csv.gz analysis [#84](https://github.com/etalab/udata-hydra/pull/84)
 
 
 ## 1.0.1 (2023-01-04)